Add files via upload
Here is another solution, instead of choosing "i" and "j", here I chose "start" and "end" from the beginning and the ending of the string to search if the first number is the same as the last number. Most of the code is similar to the previous one, the only difference is the Loop in Loop.